requirements.txt
PySide6>=6.5.0
pyserial>=3.5
qt-material>=2.14
tomli>=2.0.1
gitpython>=3.1.30
psutil>=5.9.0

# Type checking and documentation
mypy>=1.5.0
types-PySide6>=6.5.0
types-pyserial>=3.5.0
sphinx>=7.0.0
sphinx-rtd-theme>=1.3.0
sphinx-autodoc-typehints>=1.24.0
sphinx-coverage>=1.0.0

# Development tools
pre-commit>=3.3.0
black>=23.0.0
flake8>=6.0.0
pylint>=2.17.0

# Testing dependencies
pytest>=7.4.0
pytest-qt>=4.2.0
pytest-cov>=4.1.0
pytest-mock>=3.11.0
pytest-asyncio>=0.21.0
pytest-xdist>=3.3.0
pytest-html>=3.2.0
pytest-benchmark>=4.0.0
factory-boy>=3.3.0
faker>=19.0.0
coverage>=7.3.0